file upload and download now working!

+++ controllers/admin.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 const { validationResult } = require('express-validator');
@@ -135,6 +135,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDescription;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image;
       }
       return product.save().then((result) => {
@@ -170,7 +171,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found!'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then((result) => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
